pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "triage-flow-env"
version = "1.0.0"
description = "A real-time customer support ticket triage and resolution environment for OpenEnv."
readme = "README.md"
authors = [{ name = "yashaskn8" }]
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.1",
    "openai>=1.30.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "numpy>=1.26.4",
    "PyYAML>=6.0.1",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.1.0",
    "openenv-core>=0.2.0",
]

[project.scripts]
server = "server.app:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
]
